Added validation for ANDROID_HOME environment variable and added validat...

+++ Tasks/AndroidBuild/KillAndroidEmulator.ps1
@@ -5,8 +5,22 @@ param(
 Write-Verbose "Entering script KillAndroidEmulator.ps1"
 Write-Verbose "emulatorName = $emulatorName"
 
+if ($env:ANDROID_HOME -eq $null)
+{
+    throw 'Environment variable not set: ANDROID_HOME'
+}
+
 $adbexe = $env:ANDROID_HOME + "\platform-tools\adb.exe"
+if (!(Test-Path -Path $adbexe))
+{
+    throw "File not found: $adbexe"
+}
+
 $androidbat = $env:ANDROID_HOME + "\tools\android.bat"
+if (!(Test-Path -Path $androidbat))
+{
+    throw "File not found: $androidbat"
+}
 
 # Delete emulator device.  Stop-Process is used because Wait-Job or Stop-Job hangs.
 Stop-Process -processname emulator-x86
